+++ src/hooks/customs/useFavoriteMeetings.tsx
@@ -2,53 +2,15 @@ import { useQuery } from '@tanstack/react-query';
 import { meetingService } from '@/app/(home)/favorite-meetings/meetingService';
 import { useEffect, useState } from 'react';
 import { useAuthStore } from '@/store/useAuthStore';
-import { IMeeting } from '@/types/meetingsType';
+import type { IMeeting } from '@/types/meetingsType';
 import { useFilter } from './useFilter';
 import { useLikeNotify } from './useLikeNotify';
-
-// 로컬 스토리지에서 값을 가져오는 함수
-/** utils로 변경 예정 */
-function getLocalStorageItem<T>(key: string, defaultValue: T): T {
-	try {
-		const storedValue = localStorage.getItem(key);
-		if (storedValue) {
-			return JSON.parse(storedValue) as T;
-		} else {
-			return defaultValue;
-		}
-	} catch (error) {
-		console.error(
-			`Error parsing JSON from localStorage for key "${key}":`,
-			error,
-		);
-		return defaultValue;
-	}
-}
-
-/** 로그인한 사용자 또는 guest ID를 반환하는 함수 */
-/** utils로 변경 예정 */
-function getLikerKey({
-	likeList,
-	user,
-	isLoggedIn,
-}: {
-	likeList: ILikeListJSON;
-	user: number | null;
-	isLoggedIn: boolean;
-}) {
-	return Object.keys(likeList).find((key) => {
-		if (!(user && isLoggedIn)) {
-			// 로그인되지 않은 경우 guestId를 사용
-			const guestId = getLocalStorageItem<string>('guestId', '');
-			return Number(key) === Number(guestId);
-		}
-		return Number(key) === Number(user);
-	});
-}
-
-interface ILikeListJSON {
-	[key: string]: number[] | string[];
-}
+import {
+	deleteLikeId,
+	getLikerKey,
+	getLocalStorageItem,
+} from '@/utils/localStorage';
+import type { ILikeListJSON } from '@/types/likeButtonType';
 
 export const useFavoriteMeetings = () => {
 	const { type } = useFilter();
